Filter 'msls_main_save' lets you override the save-method

+++ includes/MslsMain.php
@@ -44,39 +44,44 @@ abstract class MslsMain {
 	 * @param array $input
 	 */
 	protected function save( $object_id, $class ) {
-		$input = array(
-			$this->blogs->get_current_blog()->get_language() => (int) $object_id,
-		);
-		foreach ( $_POST as $key => $value ) {
-			if ( false !== strpos( $key, 'msls_input_' ) && !empty( $value ) ) {
-				$input[substr( $key, 11 )] = (int) $value;
-			}
+		if ( has_filter( 'msls_main_save' ) ) {
+			apply_filters( 'msls_main_save', $object_id, $class );
 		}
-		$msla      = new MslsLanguageArray( $input );
-		$options   = new $class( $object_id );
-		$language  = $this->blogs->get_current_blog()->get_language();
-		$temp      = $options->get_arr();
-		$object_id = $msla->get_val( $language );
-		if ( 0 != $object_id ) 
-			$options->save( $msla->get_arr( $language ) );
-		else
-			$options->delete();
-		$blogs = $this->blogs->get();
-		foreach ( $blogs as $blog ) {
-			switch_to_blog( $blog->userblog_id );
-			$language  = $blog->get_language();
-			$object_id = $msla->get_val( $language );
-			if ( 0 != $object_id ) { 
-				$options = new $class( $object_id );
-				$options->save( $msla->get_arr( $language ) );
-			}
-			else {
-				if ( isset( $temp[$language] ) ) {
-					$options = new $class( $temp[$language] );
-					$options->delete();
+		else {
+			$input = array(
+				$this->blogs->get_current_blog()->get_language() => (int) $object_id,
+			);
+			foreach ( $_POST as $key => $value ) {
+				if ( false !== strpos( $key, 'msls_input_' ) && !empty( $value ) ) {
+					$input[substr( $key, 11 )] = (int) $value;
 				}
 			}
-			restore_current_blog();
+			$msla      = new MslsLanguageArray( $input );
+			$options   = new $class( $object_id );
+			$language  = $this->blogs->get_current_blog()->get_language();
+			$temp      = $options->get_arr();
+			$object_id = $msla->get_val( $language );
+			if ( 0 != $object_id ) 
+				$options->save( $msla->get_arr( $language ) );
+			else
+				$options->delete();
+			$blogs = $this->blogs->get();
+			foreach ( $blogs as $blog ) {
+				switch_to_blog( $blog->userblog_id );
+				$language  = $blog->get_language();
+				$object_id = $msla->get_val( $language );
+				if ( 0 != $object_id ) { 
+					$options = new $class( $object_id );
+					$options->save( $msla->get_arr( $language ) );
+				}
+				else {
+					if ( isset( $temp[$language] ) ) {
+						$options = new $class( $temp[$language] );
+						$options->delete();
+					}
+				}
+				restore_current_blog();
+			}
 		}
 	}
